+++ CHANGELOG.md
@@ -2,6 +2,11 @@
 
 ## Version 0.1 (development)
 
-- Feature A added
-- FIX: nasty bug #1729 fixed
-- add your changes here!
+> Paths used in this project are meant only for devcontainers. If you are using any other platform, kindly update the paths.
+
+- Updated enviroment.yml file
+- Created configs for `densenet`, `resnet50` & `vgg16`
+- initiallized `dvc` in project
+- added `preprocess_data.py` and `train_model.py` scripts
+- added a configuration file `data.yaml` that contains paths to all data files
+- added `data_load.py`, `loss.py` and `build_model.py` scripts